+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unrelease]
 
+### Changed
 
+- qq heartbeat ack 和 qqofficial gateway resumed 不再记录log
 
 ## [v0.2.0-alpha - 2026-06-27]